requirements.txt
Flask==1.0.2
Jinja2==2.10
MarkupSafe==1.0
Werkzeug==0.14.1
argparse==1.4.0
gunicorn==19.9.0
itsdangerous==1.1.0
requests==2.20.0
billboard.py==5.3.0
python-firebase==1.2
gspread==3.6.0
gcloud==0.18.3
google-api-core==1.21.0
google-api-python-client==1.10.0
google-auth==1.19.0
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.1
oauth2client==4.1.3
oauthlib==3.1.0
